requirements.txt
django==2.1.10
djangorestframework==3.9.4
django-cors-headers==3.0.2
django-registration-redux==2.4
django-dual-authentication==1.0
django-cron==0.5.1
django-dbbackup==3.2.0
overpass==0.6.1
easy-thumbnails==2.6.0
django-debug-toolbar
